Mark `height` of `CrystallizeImageVariant` as optional
According to the GraphQL schema, `height` is optional:

```
type ImageVariant {
  url: String!
  key: String!
  width: Int!
  height: Int
  size: Int
}
```

This means that when using something like
[GraphQL Code Generator](https://www.graphql-code-generator.com/), the generated
types will become:

```
type ImageVariant = {
  __typename?: 'ImageVariant';
  url: Scalars['String'];
  key: Scalars['String'];
  width: Scalars['Int'];
  height?: Maybe<Scalars['Int']>;
  size?: Maybe<Scalars['Int']>;
};
```

These types are incompatible as `height` is marked as optional in the generated
type.

Looking at the code, it looks like only `width` is required to calculate the
`biggestImage` variant. The resulting `height` which is used in `commonProps`
can be `undefined` as we may not have a `biggestImage` (`height: biggestImage?.height`).